backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
stable-baselines3==2.3.2
qrcode[pil]==7.4.2
python-dotenv==1.0.1
pydantic==2.7.1
alembic==1.13.1
pillow==10.3.0
torch==2.3.0
numpy==1.26.4